requirements.txt
torch>=2.0.0
ultralytics>=8.0.200
opencv-python>=4.8.0
matplotlib>=3.7.0
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0